Configurations

+++ system/app/controllers/editar-perfil-configuraciones.php
@@ -1,53 +1,78 @@
 <?php
-	
-include_once ('../../framework.php');
+	include_once ( dirname(dirname(__DIR__)) . '/framework.php');
+	$ObjMante   = new Mantenimientos();
+	$ObjEjec    = new ejecutorSQL();
 
 	$id_user 		=	$_SESSION['id_user'];
-	$id_empresa 	=	$_SESSION['id_empresa'];
+	$id_empresa 	=	$_SESSION['id_cia'];
 	$P_Tabla 		=	"usuarios";
 
-
 // Add
-if ( $_GET['add'] == 1 && $_GET['nombre'] != '') {
+if ( isset($_GET['add']) && $_GET['add'] == 1 && $_GET['nombre'] != '') {
 
-	$datosEmpresa 	=	mysql_fetch_array(mysql_query("Select * From usuarios Where id_usuario = '".$id_user."'"));
-	//echo $datosEmpresa['name_cia'];
+	$where 			= 	'usuario="'.$_GET['email'].'" and id_cia="'.$id_empresa.'"';
+	$busca 			=	$ObjMante->BuscarLoQueSea('*',PREFIX.'users',$where);
 
-    $PCLAVE			=	"AES_ENCRYPT('".htmlentities('123456')."','toga')";
-	$P_Campos 		=	'usuario,contrasena, email, nombre, apellido, id_empresa,name_cia,fecha_registro,fecha_ult_act,principal,idioma,activo,telephone,direcction,tipo_moneda';
-	$P_Valores 		=	"'".$_GET['email']."', AES_ENCRYPT('123456','toga') , '".$_GET['email']."', '".$_GET['nombre']."', '---' ,'".$_SESSION['id_empresa']."', '".$datosEmpresa['name_cia']."' , '".date("Y-m-d H:i:s")."' , '".date("Y-m-d H:i:s")."' , 0 , '".$datosEmpresa['idioma']."' , '".$_GET['estado']."', '".$_GET['telefono']."', '".$_GET['direccion']."', '".$datosEmpresa['tipo_moneda']."'";
+    //$PCLAVE			=	"AES_ENCRYPT('".htmlentities('123456')."','toga')";
+	$P_Campos 		=	'usuario,contrasena, email, nombre, id_cia,created_at,is_principal,activo,telephone,direcction,profile_photo';
+	$P_Valores 		=	"'".$_GET['email']."', AES_ENCRYPT('".$_GET['contrasena']."','toga') , '".$_GET['email']."', '".$_GET['nombre']."', '".$_SESSION['id_cia']."', NOW() , '0' , '".$_GET['estado']."', '".$_GET['telefono']."', '".$_GET['direccion']."','-'";
 	
-	$busca 			=	mysql_query("Select * From ".$P_Tabla." Where email = '".$_GET['email']."'"); // $ObjMante->BuscarLoQueSea('*' , $P_Tabla, ' codigo ='.$_POST['nombre'], 'extract', false);
-
-	if (mysql_num_rows($busca) >0 ) {
-			echo $mssg	=	'Ya existe este email.';
+	if ($busca['total'] > 0 ) {
+		echo '<div class="alert alert-danger">Este usuario ya existe.</div>';
 	} else {
-		//$sql 	=	$Objsql->insertarRegistro($P_Tabla, $P_Campos, $P_Valores);
-		$sql 	=	mysql_query("Insert into ".$P_Tabla." (".$P_Campos.") values(".$P_Valores.")") or die(mysql_error());
-		if ($sql) {
-			echo $mssg 	=	'Se ingreso el registro con éxito';
-		}
-
+		$mensaG			=	"<font face=verdana size=1.5 />Hola ".$_GET['nombre']."&nbsp;<br /><br />
+					
+		&nbsp;&nbsp;Sus datos de acceso para utilizar H&HSystem fueron creados.<br><br>
+		&nbsp;&nbsp;Esto son tus datos de acceso:<br>
+		&nbsp;&nbsp;Nombre de usuario: ".$_GET['email']."<br>			
+		&nbsp;&nbsp;Contraseña: ".$_GET['contrasena']."<br><br />
+		Bienvenido a H&HSystem<br />
+		Derechos Reservados ".date('Y')."
+		";
+		$ObjEjec->insertarRegistro(PREFIX.'users', $P_Campos, $P_Valores);
+		$mail_to_send_to = $_GET['email'];
+		$from_email 	 = $_ENV['MAIL_FROM_ADDRESS'];
+		$subject		 = "Usuario Creado";
+		$headers  = "From: " . strip_tags($from_email) . "\r\n";
+		$headers .= "Reply-To: " . strip_tags($_ENV["MAIL_USERNAME"]) . "\r\n";
+		$headers .= "BCC: ".$_ENV["MAIL_BBC"]."\r\n";
+		$headers .= "MIME-Version: 1.0\r\n";
+		$headers .= "Content-Type: text/html; charset=UTF-8\r\n";
+		$a = mail( $mail_to_send_to, $subject, $mensaG, $headers );
+		echo '<div class="alert alert-success">Se ha creado el registro con éxito. <br />Se envó un email al usuario creado, con sus datos de acceso.</div>';
 	}
+	
+	//$busca 			=	mysql_query("Select * From ".$P_Tabla." Where email = '".$_GET['email']."'"); // $ObjMante->BuscarLoQueSea('*' , $P_Tabla, ' codigo ='.$_POST['nombre'], 'extract', false);
+
+
+	// if (mysql_num_rows($busca) >0 ) {
+	// 		echo $mssg	=	'Ya existe este email.';
+	// } else {
+	// 	//$sql 	=	$Objsql->insertarRegistro($P_Tabla, $P_Campos, $P_Valores);
+	// 	$sql 	=	mysql_query("Insert into ".$P_Tabla." (".$P_Campos.") values(".$P_Valores.")") or die(mysql_error());
+	// 	if ($sql) {
+	// 		echo $mssg 	=	'Se ingreso el registro con éxito';
+	// 	}
+
+	// }
 }
 
 // Edit
-//if ($_POST['btn_edit_room']) {
-if ( $_GET['edit'] == 1 && $_GET['nombre'] != '' && $_GET['email'] !='') {
+if ( isset($_GET['edit']) &&  $_GET['edit'] == 1 && $_GET['nombre'] != '' && $_GET['email'] !='') {
 
-
-	$P_Valores 	= 	" nombre = '".$_GET['nombre']."', email = '".$_GET['email']."', fecha_ult_act = '".date("Y-m-d H:i:s")."', telephone = '".$_GET['telefono']."', direcction = '".$_GET['direccion']."', activo = '".$_GET['activo']."'";
+	$P_Valores 		= 	" nombre = '".$_GET['nombre']."', email = '".$_GET['email']."', updated_at = NOW(), telephone = '".$_GET['telefono']."', direcction = '".$_GET['direccion']."', activo = '".$_GET['activo']."'";
     if ($_GET['contrasena']!='') { $P_Valores 	.=	" , contrasena = AES_ENCRYPT(".$_GET['contrasena'].",'toga')"; }
-	
-  
-	$P_condicion=	" id_usuario = '".$_GET['id']."'";
-	//$resultOpe  =  	$Objsql->actualizarRegistro($P_Valores, $P_Tabla, $P_condicion);
-	mysql_query("Update ".$P_Tabla." set ".$P_Valores." Where ".$P_condicion."") or die(mysql_error());
-	echo $mssg 		=	'Se actualizo el registro con éxito';
+	$P_condicion 	=	" id_usuario = '".$_GET['id']."'";
+	$upd 			=	$ObjEjec->actualizarRegistro($P_Valores, PREFIX.'users', $P_condicion);
+  	if ($upd) {
+		echo 'ok';
+	} else {
+		echo 'error';
+	}
 }
 
 // Delete 
-if ( $_GET['delete'] == 1 ) {
+if ( isset($_GET['delete']) && $_GET['delete'] == 1 ) {
 	mysql_query("Delete from zz_permisos Where id_usuario = '".$_GET['id']."'") or die(mysql_error());
 	mysql_query("Delete from ".$P_Tabla." Where id_usuario = '".$_GET['id']."'") or die(mysql_error());
 	echo $mssg 		=	'Se elimino el registro con éxito';
